commented out blocks of test code that directly affect values in setboundaries.  remove range 56 by 28 removes all of the content from the first loop except planets being put into the isvisable list.
remove range 119,9   or 91, 6 if you have the first remove active removes the section in the bottom loop that calls normalised position.

Both patches work patch 1 makes the map massive, but does not render the planets as offset and mapsize variable need to be adjusted as well.

path 2 works but crashes the game, it would either need to be added to each method that calls it or preferably added to the method itself.